feat: added support for both omni directional and differential robot
