--- urdf/omni_robot.xacro
@@ -1,53 +1,0 @@
-<?xml version="1.0"?>
-<robot name="omni_robot" xmlns:xacro="http://wiki.ros.org/xacro">
-
-  <xacro:include filename="$(find haruto_description)/urdf/base.xacro"/>
-  <xacro:include filename="$(find haruto_description)/urdf/pillar.xacro"/>
-  <xacro:include filename="$(find haruto_description)/urdf/motor.xacro"/>
-  <xacro:include filename="$(find haruto_description)/urdf/lidar.xacro"/>
-  <xacro:include filename="$(find haruto_description)/urdf/omni_wheel.xacro"/>
-
-  <link name="world_link" />
-
-  <xacro:robot_base parent="world_link" child="first_base_link" xyz="0 0 0.075" rpy="0 0 0" />
-  <xacro:robot_base parent="first_base_link" child="second_base_link" xyz="0 0 0.06" rpy="0 0 0" />
-  <xacro:robot_base parent="second_base_link" child="third_base_link" xyz="0 0 0.06" rpy="0 0 0" />
-
-  <xacro:robot_pillar parent="first_base_link" child="first_front_left_pillar_link" xyz="0.11 0.065 0.03" rpy="0 0 0" />
-  <xacro:robot_pillar parent="first_base_link" child="first_front_right_pillar_link" xyz="0.11 -0.065 0.03" rpy="0 0 0" />
-  <xacro:robot_pillar parent="first_base_link" child="first_back_left_pillar_link" xyz="-0.11 0.065 0.03" rpy="0 0 0" />
-  <xacro:robot_pillar parent="first_base_link" child="first_back_right_pillar_link" xyz="-0.11 -0.065 0.03" rpy="0 0 0" />
-
-  <xacro:robot_pillar parent="second_base_link" child="second_front_left_pillar_link" xyz="0.11 0.065 0.03" rpy="0 0 0" />
-  <xacro:robot_pillar parent="second_base_link" child="second_front_right_pillar_link" xyz="0.11 -0.065 0.03" rpy="0 0 0" />
-  <xacro:robot_pillar parent="second_base_link" child="second_back_left_pillar_link" xyz="-0.11 0.065 0.03" rpy="0 0 0" />
-  <xacro:robot_pillar parent="second_base_link" child="second_back_right_pillar_link" xyz="-0.11 -0.065 0.03" rpy="0 0 0" />
-
-  <xacro:robot_motor parent="first_base_link" child="front_left_motor_link" xyz="0.10 0.040 -0.025" rpy="0 0 0" />
-  <xacro:robot_motor parent="first_base_link" child="front_right_motor_link" xyz="0.10 -0.040 -0.025" rpy="0 0 0" />
-  <xacro:robot_motor parent="first_base_link" child="back_left_motor_link" xyz="-0.10 0.040 -0.025" rpy="0 0 0" />
-  <xacro:robot_motor parent="first_base_link" child="back_right_motor_link" xyz="-0.10 -0.040 -0.025" rpy="0 0 0" />
-
-  <xacro:omni_wheel parent="front_left_motor_link" child="front_left_wheel_link" xyz="0.0 0.060 -0.015" rpy="1.57 0 0" />
-  <xacro:omni_wheel parent="front_right_motor_link" child="front_right_wheel_link" xyz="0.0 -0.060 -0.015" rpy="1.57 0 0" />
-  <xacro:omni_wheel parent="back_left_motor_link" child="back_left_wheel_link" xyz="0.0 0.060 -0.015" rpy="1.57 0 0" />
-  <xacro:omni_wheel parent="back_right_motor_link" child="back_right_wheel_link" xyz="0.0 -0.060 -0.015" rpy="1.57 0 0" />
-
-  <xacro:lidar parent="third_base_link" child="lidar_link" xyz="0.04 0.0 0.035" rpy="0 0 0" />
-
-  <gazebo>
-    <plugin name="object_controller" filename="libgazebo_ros_planar_move.so">
-      <ros>
-        <argument>odom:=odom/unfiltered</argument>
-      </ros>
-      <command_topic>cmd_vel</command_topic>
-      <odometry_topic>odom/unfiltered</odometry_topic>
-      <odometry_frame>odom</odometry_frame>
-      <odometry_rate>20.0</odometry_rate>
-      <robot_base_frame>first_base_link</robot_base_frame>
-      <publish_odom>true</publish_odom>
-      <publish_odom_tf>true</publish_odom_tf>
-    </plugin>
-  </gazebo>
-
-</robot>
