--- urdf/omni_wheel.xacro
@@ -1,36 +1,0 @@
-<?xml version="1.0"?>
-
-<robot xmlns:xacro="http://wiki.ros.org/xacro">
-
-  <xacro:macro name="omni_wheel" params="parent child xyz rpy">
-
-    <link name="${child}">
-      <visual>
-        <geometry>
-          <cylinder radius="0.0375" length="0.04"/>
-        </geometry>
-        <material name="blue">
-          <color rgba="0.0 0.0 1 1" />
-        </material>
-      </visual>
-      <collision>
-        <geometry>
-          <cylinder radius="0.0375" length="0.04"/>
-        </geometry>
-      </collision>
-      <inertial>
-        <mass value="1.0" />
-        <inertia ixx="0.015" ixy="0" ixz="0" iyy="0.0375" iyz="0.0" izz="0.0375" />
-      </inertial>
-    </link>
-
-    <joint name="${parent}_to_${child}_joint" type="continuous">
-      <axis xyz="0.0 0.0 1"/>
-      <origin xyz="${xyz}" rpy="${rpy}" />
-      <parent link="${parent}" />
-      <child link="${child}" />
-    </joint>
-
-  </xacro:macro>
-
-</robot>
