+++ urdf/robot.xacro
@@ -1,0 +1,89 @@
+<?xml version="1.0"?>
+<robot name="robot" xmlns:xacro="http://wiki.ros.org/xacro">
+
+  <xacro:arg name="robot_type" default="omni"/>
+
+  <xacro:property name="type" value="$(arg robot_type)" />
+
+  <xacro:include filename="$(find haruto_description)/urdf/base.xacro"/>
+  <xacro:include filename="$(find haruto_description)/urdf/pillar.xacro"/>
+  <xacro:include filename="$(find haruto_description)/urdf/motor.xacro"/>
+  <xacro:include filename="$(find haruto_description)/urdf/lidar.xacro"/>
+  <xacro:include filename="$(find haruto_description)/urdf/wheel.xacro"/>
+
+  <link name="world_link" />
+
+  <xacro:robot_base parent="world_link" child="first_base_link" xyz="0 0 0.075" rpy="0 0 0" />
+  <xacro:robot_base parent="first_base_link" child="second_base_link" xyz="0 0 0.06" rpy="0 0 0" />
+  <xacro:robot_base parent="second_base_link" child="third_base_link" xyz="0 0 0.06" rpy="0 0 0" />
+
+  <xacro:robot_pillar parent="first_base_link" child="first_front_left_pillar_link" xyz="0.11 0.065 0.03" rpy="0 0 0" />
+  <xacro:robot_pillar parent="first_base_link" child="first_front_right_pillar_link" xyz="0.11 -0.065 0.03" rpy="0 0 0" />
+  <xacro:robot_pillar parent="first_base_link" child="first_back_left_pillar_link" xyz="-0.11 0.065 0.03" rpy="0 0 0" />
+  <xacro:robot_pillar parent="first_base_link" child="first_back_right_pillar_link" xyz="-0.11 -0.065 0.03" rpy="0 0 0" />
+
+  <xacro:robot_pillar parent="second_base_link" child="second_front_left_pillar_link" xyz="0.11 0.065 0.03" rpy="0 0 0" />
+  <xacro:robot_pillar parent="second_base_link" child="second_front_right_pillar_link" xyz="0.11 -0.065 0.03" rpy="0 0 0" />
+  <xacro:robot_pillar parent="second_base_link" child="second_back_left_pillar_link" xyz="-0.11 0.065 0.03" rpy="0 0 0" />
+  <xacro:robot_pillar parent="second_base_link" child="second_back_right_pillar_link" xyz="-0.11 -0.065 0.03" rpy="0 0 0" />
+
+  <xacro:robot_motor parent="first_base_link" child="front_left_motor_link" xyz="0.10 0.040 -0.025" rpy="0 0 0" />
+  <xacro:robot_motor parent="first_base_link" child="front_right_motor_link" xyz="0.10 -0.040 -0.025" rpy="0 0 0" />
+  <xacro:robot_motor parent="first_base_link" child="back_left_motor_link" xyz="-0.10 0.040 -0.025" rpy="0 0 0" />
+  <xacro:robot_motor parent="first_base_link" child="back_right_motor_link" xyz="-0.10 -0.040 -0.025" rpy="0 0 0" />
+
+  <xacro:wheel parent="front_left_motor_link" child="front_left_wheel_link" xyz="0.0 0.060 -0.015" rpy="1.57 0 0" robot_type="$(arg robot_type)" />
+  <xacro:wheel parent="front_right_motor_link" child="front_right_wheel_link" xyz="0.0 -0.060 -0.015" rpy="1.57 0 0" robot_type="$(arg robot_type)" />
+  <xacro:wheel parent="back_left_motor_link" child="back_left_wheel_link" xyz="0.0 0.060 -0.015" rpy="1.57 0 0" robot_type="$(arg robot_type)" />
+  <xacro:wheel parent="back_right_motor_link" child="back_right_wheel_link" xyz="0.0 -0.060 -0.015" rpy="1.57 0 0" robot_type="$(arg robot_type)" />
+
+  <xacro:lidar parent="third_base_link" child="lidar_link" xyz="0.04 0.0 0.035" rpy="0 0 0" />
+
+  <xacro:if value="${type == 'omni'}">
+    <gazebo>
+      <plugin name="object_controller" filename="libgazebo_ros_planar_move.so">
+        <ros>
+          <argument>odom:=odom/unfiltered</argument>
+        </ros>
+        <command_topic>cmd_vel</command_topic>
+        <odometry_topic>odom/unfiltered</odometry_topic>
+        <odometry_frame>odom</odometry_frame>
+        <odometry_rate>20.0</odometry_rate>
+        <robot_base_frame>world_link</robot_base_frame>
+        <publish_odom>true</publish_odom>
+        <publish_odom_tf>true</publish_odom_tf>
+      </plugin>
+    </gazebo>
+  </xacro:if>
+
+  <xacro:if value="${type == 'diff'}">
+   <gazebo>
+      <plugin name="skid_steer_drive" filename="libgazebo_ros_diff_drive.so">
+
+         <num_wheel_pairs>2</num_wheel_pairs>
+         
+        <left_joint>front_left_motor_link_to_front_left_wheel_link_joint</left_joint>
+        <right_joint>front_right_motor_link_to_front_right_wheel_link_joint</right_joint>
+        <left_joint>back_left_motor_link_to_back_left_wheel_link_joint</left_joint>
+        <right_joint>back_right_motor_link_to_back_right_wheel_link_joint</right_joint>
+
+        <wheel_separation>0.215</wheel_separation>
+        <wheel_separation>0.215</wheel_separation>
+
+        <wheel_diameter>0.035</wheel_diameter>
+        <wheel_diameter>0.035</wheel_diameter>
+
+        <max_wheel_torque>200</max_wheel_torque>
+        <max_wheel_acceleration>10.0</max_wheel_acceleration>
+
+        <odometry_frame>odom</odometry_frame>
+        <robot_base_frame>world_link</robot_base_frame>
+
+        <publish_odom>true</publish_odom> 
+        <publish_odom_tf>true</publish_odom_tf> 
+        <publish_wheel_tf>true</publish_wheel_tf> 
+      </plugin>
+     </gazebo>
+  </xacro:if>
+
+</robot>
